oidc: Add top-level `oidc` module
